+++ lib/src/point.dart
@@ -1,0 +1,11 @@
+/// 2D integer pixel coordinate.
+class Point {
+  /// X coordinate in pixels
+  final int x;
+
+  /// Y coordinate in pixels
+  final int y;
+
+  /// Creates a 2D pixel coordinate at position ([x], [y]).
+  Point(this.x, this.y);
+}